+++ execution/engine.go
@@ -351,15 +351,6 @@ func (e *Engine) submitMarket(ctx context.Context, marketConfig *types.Market) e
 			logging.AssetID(asset))
 	}
 
-	// @todo no need to set it
-	// set a fake tick size to the continuous trading if it's continuous
-	switch tmod := marketConfig.TradingModeConfig.(type) {
-	case *types.MarketContinuous:
-		tmod.Continuous.TickSize = e.getFakeTickSize(marketConfig.DecimalPlaces)
-	case *types.MarketDiscrete:
-		tmod.Discrete.TickSize = e.getFakeTickSize(marketConfig.DecimalPlaces)
-	}
-
 	// create market auction state
 	mas := monitor.NewAuctionState(marketConfig, now)
 	ad, err := e.assets.Get(asset)