+++ CHANGELOG.md
@@ -64,6 +64,9 @@
 - [4680](https://github.com/vegaprotocol/vega/issues/4680) - Add `totalTokenSupplyStake` to the snapshots
 - [4645](https://github.com/vegaprotocol/vega/pull/4645) - Add transfers snapshots
 - [4595](https://github.com/vegaprotocol/vega/pull/4595) - Add internal oracle supplying vega time data for time-triggered events
+- [4681](https://github.com/vegaprotocol/vega/issues/4681) - Remove tick size from market 
+- [4698](https://github.com/vegaprotocol/vega/issues/4698) - Remove maturity field from future
+- [4699](https://github.com/vegaprotocol/vega/issues/4699) - Remove trading mode one off from market proposal
 
 ### 🐛 Fixes
 - [4521](https://github.com/vegaprotocol/vega/pull/4521) - Better error when trying to use the null-blockchain with an ERC20 asset